% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_catches.R
\name{plot_catches}
\alias{plot_catches}
\title{Plot catches for each species by gear}
\usage{
plot_catches(year, species, area)
}
\arguments{
\item{year}{latest year to include}

\item{species}{species group code, e.g. "DUSK" found where}

\item{area}{sample area "GOA", "AI" or "BS" (or combos)}
}
\value{

}
\description{
Plot catches for each species by gear
}
\examples{
plot_catches(year = 2022,species = "DUSK",area = "GOA")
}
