% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_total_catch.R
\name{plot_total_catch}
\alias{plot_total_catch}
\title{Plot catches for each species by gear}
\usage{
plot_total_catch(year, species, area, db = akfin)
}
\arguments{
\item{year}{latest year to include}

\item{species}{species group code, e.g. "DUSK" found where}

\item{area}{sample area "GOA", "AI" or "BS" (or combos)}

\item{db}{database to use}
}
\value{

}
\description{
Plot catches for each species by gear
}
\examples{
plot_total_catch(year = 2022,species = "DUSK",area = "GOA",db = akfin)
}
